test: relocate api spec test
